backend/requirements.txt
fastapi
yfinance
pydantic
sqlalchemy
apscheduler
pandas_market_calendars
pytest
httpx
uvicorn
discord.py
python-dotenv
